requirements.txt
pandas>=1.0.0
numpy>=1.20.0
matplotlib>=3.0.0
reportlab>=3.5.0
openai>=1.81.0
python-dotenv>=0.21.0